handle no url no schema file with useLocal true
Previously, server could get into an undefined state when useLocalSchema = true and no schema source was specified.
The configuration tools will throw an error if that scenario is discovered.
If useLocalSchema is false and no url is provided and the apollo registry has a url, then that url will be used